return login and signup tokens

+++ src/resolvers/document.resolver.ts
@@ -6,8 +6,8 @@ import { ObjectId } from 'bson';
 const documentResolver = {
   Mutation: {
     async createDocument(root: any, args: any, context: any) {
-      console.log(context);
       if (!context.user) return [];
+      if(context.user.signUp) throw new Error('Problem with token, not auth token');
       const userFinded =await UserMong.findOne({ email: context.user.email });
       const documentMong_new = new DocumentModel({
         id: ObjectId,
@@ -24,16 +24,24 @@ const documentResolver = {
     },
     deleteDocument(root: any, args: any, context: any) {
       if (!context.user) return [];
+      if(context.user.signUp) throw new Error('Problem with token, not auth token');
       //return DocumentModelController.deleteDocument(args.tittle);
       return DocumentModel.deleteOne({tittle: args.tittle, type: args.type});
     },
     async updateDocument(root: any, args: any, context: any) {
       if (!context.user) return [];
-      const existDocument = await DocumentModel.findOne(args.tittle);
+      if(context.user.signUp) throw new Error('Problem with token, not auth token');
+      const existDocument = await DocumentModel.findOne({tittle: args.tittle});
       if (existDocument) {
         //delete tempUser.id;
-        //return DocumentModelController.updateDocument(existDocument, {args});
-        return DocumentModel.updateOne({ _id: existDocument._id }, { $set: args });
+        //return DocumentModelController.updateDocument(existDocument._id, {args});
+        return DocumentModel.findOneAndUpdate({ _id: existDocument._id }, { $set: args }, {new: true});/*,(err, doc)=>{
+          if(err){
+            return new Error ('Updating error');
+          }
+          console.log(doc.tittle);
+          return doc;
+        });*/
       } else {
         return new Error('Document doesnt exist');
       }
@@ -43,6 +51,7 @@ const documentResolver = {
   Query: {
     async documents(root: any, args: any, context: any) {
       if (!context.user) return [];
+      if(context.user.signUp) throw new Error('Problem with token, not auth token');
       const userFinded =await UserMong.findOne({ email: context.user.email });
       //return DocumentModelController.findAllDocuments(userFinded._id);
       return DocumentModel.find({user: userFinded._id});
